Partial rewrite of build process
Switching to an Arch based build process for easier maintenance and
using specific runners for tasks where that might be necessary.

Right now, parent/child pipelines for somewhat parallel builds for
different architectures / semi-independent pak-builds don't seem to
properly work yet due to `trigger:` not being recognized when it should,
so that's not yet implemented.